moved writeable block
