+++ tempodb/wal/block.go
@@ -6,17 +6,7 @@ import (
 	"sync"
 
 	"github.com/grafana/tempo/tempodb/encoding"
-	"github.com/grafana/tempo/tempodb/encoding/bloom"
 )
-
-type WriteableBlock interface {
-	BlockMeta() *encoding.BlockMeta
-	BloomFilter() *bloom.ShardedBloomFilter
-	Records() []*encoding.Record
-	ObjectFilePath() string
-
-	Flushed() error
-}
 
 type block struct {
 	meta     *encoding.BlockMeta
